Use OffscreenCanvas for drawing the favicon if possible
The advantage of using an OffscreenCanvas is that it is decoupled from
the DOM because it's rendered off-screen. This eliminates some
synchronisation overhead and should be more responsive than a regular
canvas.

The necessary types for using OffscreenCanvas properly aren't bundled
into TypeScript yet so add @types/web as a dependency too.