Draw the marquee using requestAnimationFrame
The previous setInterval method meant consumers had to manually specify
the desired framerate. RequestAnimationFrame (raf) instead uses the
monitor's native framerate, making the animation smooth on all screens,
not just the lowest common denominator, as is the case when specifying
the interval manually.

Since the marquee scrolls a lot faster now, increase the default value
for the size and step parameters. The increased size is no longer a
problem after fixing the resource leak in commit ea464f36.